refactor(domain): establish hexagonal domain models, ports, and service container
