+++ src/domain/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod credential;
+pub mod status_list;
